added unitTests summary (unfinished)
