+++ src/modules/test/unitTests/test.h
@@ -1,0 +1,8 @@
+#ifndef TEST_TESTDLL
+#define TEST_TESTDLL
+
+#include <stdint.h>
+
+extern uint16_t testTestDll();
+
+#endif
