+++ src/unitTestSystem/unitTest.h
@@ -5,6 +5,8 @@
 #include <stdbool.h>
 #include <stdio.h>
 
+size_t (*runFuncts[2])();
+
 extern void iterateTests(char* [], bool (*[])(), uint16_t);
 
 #endif
